Fixed methods

+++ src/main/java/ru/mail/polis/homework/simple/DoubleTask.java
@@ -36,8 +36,8 @@ public class DoubleTask {
      * Пример: (0, 0, 3, 4) -> 5.0
      */
     public static float length(double x1, double y1, double x2, double y2) {
-        float cathet1 = (float) Math.abs(x1 - x2);
-        float cathet2 = (float) Math.abs(y1 - y2);
+        double cathet1 = Math.abs(x1 - x2);
+        double cathet2 = Math.abs(y1 - y2);
 
         return (float) Math.sqrt(Math.pow(cathet1, 2) + Math.pow(cathet2, 2));
     }
